Outdated browser message using http://outdatedbrowser.com
tests if the browser supports the font-kerning CSS property (see https://caniuse.com/#search=font-kerning) and displays a nice MGS-blue message at the top of the page if not. This appears repetitively, and does not store a cookie to stop it from reappearing - it is forceful on purpose to ensure that we don't cause disappointment to users with older browsers.